+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/darts/RotDart.java
@@ -54,7 +54,7 @@ public class RotDart extends TippedDart {
 	}
 	
 	@Override
-	public float durabilityPerUse(int level) {
+	public float durabilityPerUse(float level) {
 		return MAX_DURABILITY/5f; //always 5 uses
 	}
 }
