+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/darts/TippedDart.java
@@ -172,7 +172,7 @@ public abstract class TippedDart extends Dart {
 	}
 
 	@Override
-	public float durabilityPerUse(int level) {
+	public float durabilityPerUse(float level) {
 		float use = super.durabilityPerUse(level);
 
 		if (Dungeon.hero != null) {
